Updated to 1.19.2 Added grave compass

+++ src/main/java/com/b1n_ry/yigd/block/entity/GraveBlockEntity.java
@@ -74,7 +74,7 @@ public class GraveBlockEntity extends BlockEntity {
         tag.putInt("StoredXp", storedXp);
         tag.put("Items", Inventories.writeNbt(new NbtCompound(), this.storedInventory, true));
         tag.putInt("ItemCount", this.storedInventory.size());
-        tag.putLong("age", this.age);
+        tag.putInt("age", this.age);
         tag.put("replaceState", NbtHelper.fromBlockState(this.previousState));
         tag.putBoolean("claimed", this.claimed);
         tag.putUuid("graveId", this.graveId);
